--- src/librand/NOTES.txt
@@ -1,4 +1,0 @@
-LIBRAND
-
-2000-12-29, David A. Morano
-
