--- src/librand/original/copyright.txt
@@ -1,1 +1,0 @@
-/* Copyright © 2017 David A­D­ Morano.  All rights reserved. */
