+++ src/libuc/randomvar.h
@@ -12,9 +12,6 @@
 
 
 #include	<envstandards.h>	/* MUST be first to configure */
-#include	<sys/types.h>
-#include	<sys/param.h>
-#include	<time.h>
 #include	<usystem.h>
 #include	<stdintx.h>
